ta quase

+++ aula_13_08/src/main/java/org/libertas/PessoaDAO.java
@@ -2,8 +2,11 @@ package org.libertas;
 
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
+import java.sql.SQLException;
 import java.sql.Statement;
 import java.util.LinkedList;
+
+import com.mysql.cj.protocol.Resultset;
 
 public class PessoaDAO {
 //	private static LinkedList<Pessoa> lista = new LinkedList<Pessoa>();
@@ -17,7 +20,7 @@ public class PessoaDAO {
 	}
 	
 	
-	public void inserir(Pessoa p) {
+	public boolean inserir(Pessoa p) {
 //		lista.add(p);
 		Conexao conn = new Conexao();
 		
@@ -31,11 +34,14 @@ public class PessoaDAO {
 			prep.setString(5, p.getEndereco());
 			prep.setString(6, p.getCep());
 			prep.execute();
-;		} catch (Exception e) {
+			conn.desconecta();
+			return true;
+			} catch (Exception e) {
 			// TODO: handle exception
 			e.printStackTrace();
+			conn.desconecta();
+			return false;
 		}
-		conn.desconecta();
 	}
 	
 	public void alterar(Pessoa p) {
@@ -58,20 +64,33 @@ public class PessoaDAO {
 		conn.desconecta();
 	}
 	
-	public void excluir(Pessoa p) {
+	public boolean excluir(Pessoa p){
 		Conexao conn = new Conexao();
-		try {
-			String sql = "DELETE FROM pessoa WHERE idPessoa = ?";
-			PreparedStatement prep = conn.getConnection().prepareStatement(sql);
-			prep.setInt(1, p.getIdPessoa());
-			prep.execute();
+			try {
+				String sql = "SELECT * FROM pessoa WHERE idPessoa = ?";
+				PreparedStatement prep = conn.getConnection().prepareStatement(sql);
+				prep.setInt(1, p.getIdPessoa());
+				ResultSet res = prep.executeQuery();
+				if (res.next() == false ) {
+					conn.desconecta();
+					return false;
+				}
+				
+				
+				String sql1 = "DELETE FROM pessoa WHERE idPessoa = ?";
+				PreparedStatement prep1 = conn.getConnection().prepareStatement(sql1);
+				prep1.setInt(1, p.getIdPessoa());
+				prep1.execute();
+				conn.desconecta();
+				return true;
 			} catch (Exception e) {
-			// TODO: handle exception
+				// TODO: handle exception
 			e.printStackTrace();
-		}
-		conn.desconecta();
+			conn.desconecta();
+			}
+		return false;
 	}
-
+		
 	public LinkedList<Pessoa> listar() {
 		LinkedList<Pessoa> lista = new LinkedList<Pessoa>();
 		Conexao conn = new Conexao();
@@ -104,21 +123,19 @@ public class PessoaDAO {
 			String sql = "SELECT * FROM pessoa WHERE idPessoa = " + id;
 			Statement sta = conn.getConnection().createStatement();
 			ResultSet res = sta.executeQuery(sql);
-			if (res.next()) {
-				p.setIdPessoa(res.getInt("idPessoa"));
-				p.setNome(res.getString("nome"));
-				p.setTelefone(res.getString("telefone"));
-				p.setEmail(res.getString("email"));
-				p.setCidade(res.getString("cidade"));
-				p.setEndereco(res.getString("endereco"));
-				p.setCep(res.getString("cep"));
-			}
+			p.setIdPessoa(res.getInt("idPessoa"));
+			p.setNome(res.getString("nome"));
+			p.setTelefone(res.getString("telefone"));
+			p.setEmail(res.getString("email"));
+			p.setCidade(res.getString("cidade"));
+			p.setEndereco(res.getString("endereco"));
+			p.setCep(res.getString("cep"));
 		} catch (Exception e) {
 			// TODO: handle exception
 			e.printStackTrace();
-			return p;
 		}
 		conn.desconecta();
+		return p;
 	}
 
 }
